Added LabelMe integration and tracking updates

+++ sam2/sam2_video_predictor.py
@@ -778,7 +778,8 @@ class SAM2VideoPredictor(SAM2Base):
         storage_device = inference_state["storage_device"]
         maskmem_features = current_out["maskmem_features"]
         if maskmem_features is not None:
-            maskmem_features = maskmem_features.to(torch.bfloat16)
+            # Always use float32 instead of bfloat16 to avoid dtype mismatches
+            maskmem_features = maskmem_features.to(torch.float32)
             maskmem_features = maskmem_features.to(storage_device, non_blocking=True)
         pred_masks_gpu = current_out["pred_masks"]
         # potentially fill holes in the predicted masks
@@ -830,7 +831,8 @@ class SAM2VideoPredictor(SAM2Base):
 
         # optionally offload the output to CPU memory to save GPU space
         storage_device = inference_state["storage_device"]
-        maskmem_features = maskmem_features.to(torch.bfloat16)
+        # Always use float32 instead of bfloat16 to avoid dtype mismatches
+        maskmem_features = maskmem_features.to(torch.float32)
         maskmem_features = maskmem_features.to(storage_device, non_blocking=True)
         # "maskmem_pos_enc" is the same across frames, so we only need to store one copy of it
         maskmem_pos_enc = self._get_maskmem_pos_enc(
